fix: PK 컬럼 Long 타입으로 전환

+++ src/main/kotlin/com/hyungjun/todolist/board/core/domain/BoardUser.kt
@@ -1,8 +1,8 @@
 package com.hyungjun.todolist.board.core.domain
 
 class BoardUser(
-	val id: Int = 0,
-	val userId: Int,
-	val boardId: Int,
+	val id: Long = 0,
+	val userId: Long,
+	val boardId: Long,
 	var role: BoardUserRole
 )
